third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-padding.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script src="parsing-test-utils.js"></script>
<script>
description('Test parsing of the CSS shape-padding property.');

// The test functions and globals used here are defined parsing-test-utils.js.

validShapeLengths.forEach(function(value, i, a) {
    testShapeSpecifiedProperty("shape-padding", value, value);
});

testShapeSpecifiedProperty("shape-padding", "0", "0px");

invalidShapeLengths.forEach(function(value, i, a) {
    testShapeSpecifiedProperty("shape-padding", value, "");
});

applyToEachArglist(
    testShapeComputedProperty,
    [// [property, value, expectedValue]
     ["shape-padding", "0", "0px"],
     ["shape-padding", "1px", "1px"],
     ["shape-padding", "-5em", "0px"],
     ["shape-padding", "identifier", "0px"],
     ["shape-padding", "\'string\'", "0px"]]
);

applyToEachArglist(
    testNotInheritedShapeChildProperty,
    [// [property, parentValue, childValue, expectedChildValue]
     ["shape-padding", "0", "0", "0px"],
     ["shape-padding", "0", "1px", "1px"],
     ["shape-padding", "1px", "-1em", "0px"],
     ["shape-padding", "2px", "1px", "1px"]]
);
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
