third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-padding-expected.txt

Test parsing of the CSS shape-padding property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("shape-padding", "1.5ex") is "1.5ex"
PASS getCSSText("shape-padding", "2em") is "2em"
PASS getCSSText("shape-padding", "2.5in") is "2.5in"
PASS getCSSText("shape-padding", "3cm") is "3cm"
PASS getCSSText("shape-padding", "3.5mm") is "3.5mm"
PASS getCSSText("shape-padding", "4pt") is "4pt"
PASS getCSSText("shape-padding", "4.5pc") is "4.5pc"
PASS getCSSText("shape-padding", "5px") is "5px"
PASS getCSSText("shape-padding", "0") is "0px"
PASS getCSSText("shape-padding", "-5px") is ""
PASS getCSSText("shape-padding", "auto") is ""
PASS getCSSText("shape-padding", "120%") is ""
PASS getCSSText("shape-padding", "'string'") is ""
PASS getComputedStyleValue("shape-padding", "0") is "0px"
PASS getComputedStyleValue("shape-padding", "1px") is "1px"
PASS getComputedStyleValue("shape-padding", "-5em") is "0px"
PASS getComputedStyleValue("shape-padding", "identifier") is "0px"
PASS getComputedStyleValue("shape-padding", "'string'") is "0px"
PASS getChildComputedStyle("shape-padding", "0", "0") is "0px"
PASS getChildComputedStyle("shape-padding", "0", "1px") is "1px"
PASS getChildComputedStyle("shape-padding", "1px", "-1em") is "0px"
PASS getChildComputedStyle("shape-padding", "2px", "1px") is "1px"
PASS successfullyParsed is true

TEST COMPLETE
